Refactor macros for backwards compatibility
