+++ Cargo.toml
@@ -12,7 +12,7 @@ name = "rust_decimal"
 readme = "./README.md"
 repository = "https://github.com/paupino/rust-decimal"
 rust-version = "1.67.1"
-version = "1.36.0"
+version = "1.37.0"
 
 [package.metadata.docs.rs]
 all-features = true
@@ -30,6 +30,7 @@ postgres-types = { default-features = false, optional = true, version = "0.2" }
 proptest = { default-features = false, optional = true, features = ["std"], version = "1.0" }
 rand = { default-features = false, optional = true, version = "0.8" }
 rand-0_9 = { default-features = false, optional = true, package = "rand", version = "0.9" }
+rust_decimal_macros = { path = "macros", default-features = false, optional = true }
 rkyv = { default-features = false, features = ["size_32", "std"], optional = true, version = "0.7.42" }
 rocket = { default-features = false, optional = true, version = "0.5.0-rc.3" }
 serde = { default-features = false, optional = true, version = "1.0" }
@@ -65,7 +66,7 @@ db-diesel2-postgres = ["db-diesel-postgres"]
 db-postgres = ["dep:bytes", "dep:postgres-types", "std"]
 db-tokio-postgres = ["dep:bytes", "dep:postgres-types", "std"]
 legacy-ops = []
-macro = []
+macros = ["dep:rust_decimal_macros"]
 maths = []
 maths-nopanic = ["maths"]
 ndarray = ["dep:ndarray"]
