+++ macros/tests/invalid/large_scale.stderr
@@ -4,4 +4,4 @@ error: proc macro panicked
 4 |     let _ = dec!(0.00000_00000_00000_00000_00000_0001);
   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
-  = help: message: Number has a high precision that can not be represented.
+  = help: message: Scale exceeds the maximum precision allowed: 29 > 28.
