+++ macros/tests/invalid/too_small.stderr
@@ -4,4 +4,4 @@ error: proc macro panicked
 4 |     let _ = dec!(-79_228_162_514_264_337_593_543_950_336);
   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
-  = help: message: Invalid decimal: overflow from too many digits
+  = help: message: Number less than minimum value that can be represented.
