+++ src/lib.rs
@@ -61,13 +61,13 @@ pub use maths::MathematicalOps;
 
 /// A convenience module appropriate for glob imports (`use rust_decimal::prelude::*;`).
 pub mod prelude {
-    #[cfg(feature = "macro")]
-    pub use crate::dec;
     #[cfg(feature = "maths")]
     pub use crate::maths::MathematicalOps;
     pub use crate::{Decimal, RoundingStrategy};
     pub use core::str::FromStr;
     pub use num_traits::{FromPrimitive, One, Signed, ToPrimitive, Zero};
+    #[cfg(feature = "macros")]
+    pub use rust_decimal_macros::dec;
 }
 
 #[cfg(feature = "diesel")]
